feat: add TaskStatusButton, and fix loading state for no re-render all tasks. Add date in each task
