+++ src/App.jsx
@@ -1,4 +1,4 @@
-import ToDo from "./components/ToDo/ToDo";
+import ToDo from "./components/toDo/ToDo";
 
 function App() {
   return <ToDo />;
